pkp/pkp-lib#4904 fixes after the code review
